+++ pkg/checker.go
@@ -1,0 +1,7 @@
+package gmon
+
+import "context"
+
+type Checker interface {
+	Check(ctx context.Context, handlers []Handler) error
+}
