+++ pkg/handler.go
@@ -1,0 +1,5 @@
+package gmon
+
+type Handler interface {
+	Handle()
+}
