--- pkg/model/entry.go
@@ -1,8 +1,0 @@
-package model
-
-import "time"
-
-type LogEntry struct {
-	Text string
-	Time time.Time
-}
